Extract coercible logic of constructors into Constructor::Coercible
