+++ lib/dry/types/rule.rb
